Move script hex parsing to primitives
Now that we use stable error types we can move the hex parsing
functions from `bitcoin` to `primitives`.

Leave the deprecated method behind in `bitcoin`.
