+++ bitcoin/src/taproot/mod.rs
@@ -1659,7 +1659,6 @@ mod test {
     use hex_unstable::DisplayHex;
 
     use super::*;
-    use crate::script::ScriptBufExt as _;
     use crate::sighash::TapSighashTag;
     use crate::{Address, KnownHrp, ScriptPubKeyBuf};
     extern crate serde_json;
